Move mint button to the header
The mint button is now displayed next to the on-chain token balance
because it doesn't increase the off-chain balance.
