+++ src/app/modules/raiden-icons/raiden-icons.module.ts
@@ -39,6 +39,7 @@ export class RaidenIconsModule {
         'help',
         'graph',
         'doublearrow',
+        'mint',
     ];
 
     constructor(
