+++ src/sass/colors.scss
@@ -10,10 +10,10 @@ $blue: #2c8f94;
 $green: #38c141;
 $orange: #ff5c00;
 
-$label-yellow: #fdd327;
+$label-yellow: #ffd41e;
 $label-red: #ff0000;
 $label-green: #1dc512;
-$bg-label-yellow: #655823;
+$bg-label-yellow: #5b4c09;
 $bg-label-red: #5a0d0d;
 $bg-label-green: #154912;
 
@@ -24,6 +24,7 @@ $text-grey: #888888;
 
 $bg-dark-grey: #121212;
 $bg-grey: #232323;
+$bg-dark-light-grey: #3e3e3e;
 $bg-light-grey: #cccccc;
 
 $light-divider: rgba(255, 255, 255, 0.1);
